pyproject.toml
[project]
name = "myshop"
version = "0.1.0"
description = ""
authors = [
    {name = "Mihhail327",email = "mixa-amr@hotmail.com"}
]
requires-python = ">=3.13,<4"
dependencies = [
    "django (>=6.0.4,<7.0.0)",
    "djangorestframework (>=3.17.1,<4.0.0)",
    "django-environ (>=0.13.0,<0.14.0)",
    "pillow (>=12.2.0,<13.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "django-debug-toolbar (>=6.3.0,<7.0.0)",
    "dj-database-url (>=2.1.0,<3.0.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
    "django-modeltranslation (>=0.20.2,<0.21.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "drf-spectacular (>=0.29.0,<0.30.0)",
    "ruff (>=0.15.11,<0.16.0)",
    "pytest (>=9.0.3,<10.0.0)",
    "graphene-django (>=3.2.3,<4.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120
target-version = "py313"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = ["E501"] # Игнорируем слишком длинные строки, если они читаемы

[tool.ruff.format]
quote-style = "single"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "core.settings"
python_files = ["tests.py", "test_*.py", "*_tests.py"]

[dependency-groups]
dev = [
    "ruff (>=0.15.11,<0.16.0)",
    "pytest (>=9.0.3,<10.0.0)",
    "pytest-django (>=4.12.0,<5.0.0)"
]
